add daemon tools
